# Improve performance for binary serialization tasks