--- a/rebar.config
+++ b/rebar.config
@@ -1,1 +1,12 @@
 {erl_opts, [debug_info]}.
+
+{profiles, [
+    {test, [
+        {deps, [
+            {proper, "1.3.0"}
+        ]},
+        {plugins, [
+            {rebar3_bench, "0.2.1"}
+        ]}
+    ]}
+]}.