--- a/src/thrift_strict_binary_codec.erl
+++ b/src/thrift_strict_binary_codec.erl
@@ -17,7 +17,7 @@
 %% under the License.
 %%
 
--module(thrift_strict_binary_membuffer_protocol).
+-module(thrift_strict_binary_codec).
 
 -compile(inline).