… README.md
… src/thrift.app.src
… src/thrift_strict_binary_membuffer_protocol.erl
… test/AST.thrift
… test/ast.sample.term
… test/ast_thrift_gen.erl
… test/bench_thrift_serialize.erl
… test/generic_codec_test.erl
… test/validation_test.erl