Очередная попытка выжать максимум из thrift runtime в задаче перекладывания трифтовых структур в набор байт и обратно.

Некоторые исправления коснулись обобщённого кода в `thrift_protocol`, однако дали не более 15% уменьшения таймингов сериализации и десериализации. Однако, учитывая что для 95% (наших) задач важны тайминги сериализации из трифтовых структур в бинари и десериализации обратно из бинарей в структуры, я взял и склеил `thrift_protocol`, `thrift_binary_protocol` и `thrift_membuffer_transport` в один модуль, попутно избавившись от тонны external calls, излишнего boxing'а, включив агрессивный inlining и обсыпав это всё ещё небольшой кучей микрооптимизаций, которые нет смысла перечислять. Подозреваю, что это близко к максимуму того, что можно достичь, не прибегая к рефакторингу генератора.

Результат бенчмарка сериализации/десериализации [случайно сгенерённого экземляра](https://github.com/rbkmoney/thrift_erlang/blob/977fae5960e4d16f30398dc45dea64a059ad3ac7/test/ast.sample.term) модели [`AST.Module`](https://github.com/rbkmoney/thrift_erlang/blob/977fae5960e4d16f30398dc45dea64a059ad3ac7/test/AST.thrift#L7) на Intel Core i7-8550U CPU @ 1.80GHz с выключенным скейлингом, по сравнению с мастером:

```
Min:                294.66ms (- 1065.05ms / 78.3%)
25 percentile:      295.81ms (- 1073.76ms / 78.4%)
Median:             297.21ms (- 1076.93ms / 78.4%)
75 percentile:      298.60ms (- 1082.48ms / 78.4%)
Max:                307.15ms (- 1195.99ms / 79.6%)
Outliers:         Lo: 0; Hi: 1; Sum: 1
Outlier variance: 4.75000e-2 (slight)
> Bootstrapped
Mean:               297.67ms (- 1085.40ms / 78.5%)
Std deviation:     2732.84μs
> Relative
Difference at 95.0 confidence
 -1085.40ms ±    13.90ms
   -78.48%  ±  1.00%
 (Student's t, pooled s = 2.17097e+7
```

```
===> Testing bench_thrift_serialize:bench_thrift_serialize()
===> Stats for wall_time
Min:                138.56ms (-  902.20ms / 86.7%)
25 percentile:      139.07ms (-  909.75ms / 86.7%)
Median:             139.41ms (-  912.28ms / 86.7%)
75 percentile:      139.89ms (-  913.68ms / 86.7%)
Max:                146.28ms (-  941.32ms / 86.6%)
Outliers:         Lo: 0; Hi: 2; Sum: 2
Outlier variance: 4.75000e-2 (slight)
> Bootstrapped
Mean:               140.09ms (-  913.07ms / 86.7%)
Std deviation:     2000.44μs
> Relative
Difference at 95.0 confidence
  -913.07ms ±  4422.77μs
   -86.70%  ±  0.42%
 (Student's t, pooled s = 6.91009e+6
```